Support dep-files placeholders in write and write_json
Summary:
Well, dep-files are quite the challenge for content-based paths.

In this case, if you have an argsfile that contains some content-based paths that are tagged for dep-files, then if something that is unused changes, then the argsfile will also change its path. That will cause a dep-file miss, when we want it to be a hit.

One way to fix that is to write the argsfile twice - once with their content-based paths, which is what we use in the command but we also mark it as unused for dep-files so it is ignored for dep-files, and then a second time with a placeholder for content-based paths. The version with a placeholder for content-based paths is used for dep-file comparisons.

Reviewed By: cjhopman

Differential Revision: D75965856

fbshipit-source-id: efc7ab8b0d0adeed76b03cf4cc3f81fafc26a7e0