possibility to store ActionSharedDirectory in ArtifactMetadata
Summary:
Why:

Further in the stack (adding incrementality to remote actions) I am adding output artifacts from previous action run as inputs for the next run of the same action. To do that I need `ActionDirectoryEntry<ActionSharedDirectory>` representing those outputs. Materializer is a best place to act as a source of truth for such outputs, but it packs `ActionSharedDirectory` into `DirectoryMetadata` (preserves only fingerprint + size) for optimization purposes. In this diff I'm adding a variant to `DirectoryMetadata` to store full `ActionSharedDirectory` as an option for some artifacts.

How:

Biggest challenge is to update persistency layer since `DirectoryMetadata` is serialized (as a part of `ArtifactMetadata`) to SQLite. In this diff I change the semantics of the row in corresponding SQLite table. Before the change each row represented one materialized artifact whether it was a directory or just a file. Now there are rows which only represent a part of/content of a materialized directory artifact. Those rows have newly introduced field `parent_path != NULL`.

I don't change any buck2 behavior in this diff since introduced `DirectoryMetadata::Full` variant is not used anywhere. I plan to apply it selectively only to outputs of flagged actions in the future.

Reviewed By: scottcao

Differential Revision: D77466349

fbshipit-source-id: 0b65499b71762d15bf3e3f283d8ab44e5d4769eb