load() static JSON files for consumption in Starlark
Summary:
When bridging the gap between `buck2` and other build systems, it can be useful
to allow a Starlark file to `load()` a static JSON file to inform the way it
generates targets.

This diff implements that possibility, by treating JSON files as Starlark
modules that simply export a single top-level `"value"` symbol that is the
parsed contents of the file.

Reviewed By: scottcao

Differential Revision: D76084442

fbshipit-source-id: 8874b0cd35c87a6625f90d9fc689a72f3fddda49