Add ClientRuntime
Summary:
The next diff finalizes the `EventsCtx` from a non-async context so requires having a tokio runtime at the top of the stack.

Finalizing the event log calls `tokio::io::AsyncWrite::write_all` which panics (on some platforms) if executed by a different tokio runtime than the one that creates it.
We also can't create the runtime earlier since creating a tokio runtime before forking also causes panics (on different platforms).

This holds a reference to a runtime in a `OnceLock` so we can initialize it during runtime and re-use it, or, initialize during finalization if it was never initialized.

Reviewed By: JakobDegen

Differential Revision: D76086836

fbshipit-source-id: f0a8bb3ee50c85dd9552db90e19dc6539c694f2d