Move event finalizing out of client ctx
Summary:
This is an intermediate step towards finalizing logging at the top of the stack.

This shouldn't capture new errors for existing commands but should cause us to start writing invocation records and errors for a few commands that don't call `command_ctx.exec`, specifically `root` and `status`.

I'm explicitly opting out logging for internal commands that are triggered by other commands (`forkserver`, `internal-test-runner` and later, `daemon`) since if these fail it should be propagated to the triggering command. I'm also not sure if these share the trace id of the parent command which could be an issue.

Reviewed By: Will-MingLun-Li

Differential Revision: D75581273

fbshipit-source-id: 0d890c134b124ae1e132e7596058645f83498eca