Add DaemonEventsCtx
Summary:
Currently file tailers for daemon stdout/stderr are held in `BuckdClient`, opened and closed by `FlushingBuckdClient` and used by `EventsCtx`.

This creates a wrapper around `EventsCtx` to hold file tailers while a connection to the daemon is active, to manage these tailers and to split methods that are valid only when the daemon is connected (`unpack_*`) and those that aren't.

This is a round about way to limit the lifetime of `tokio::time::Ticker`, which could be done with a smaller change but this seems worth (starting to) clean up.
This should make it possible to do a larger cleanup (getting rid of `BuckdClientConnector`, `FlushingBuckdClient` and client proc macros) but that isn't done in this stack.

The reason for limiting the lifetime of `Ticker` is to allow initializing `EventCtx` outside of an async context, since `Ticker::new` will panic if initialized outside of a tokio runtime.

Reviewed By: Will-MingLun-Li

Differential Revision: D75744746

fbshipit-source-id: 276aad42b29341aa31f18cb9c50783c8bed1b24f