Check ptxas output is not empty
Summary:
We've had a couple case where the cuda_cubin_lower action produced an empty output and exited successfully. Post - https://fb.workplace.com/groups/416386506377092/permalink/1434362727912793/. The cuda_cubin_lower action calls the Nvidia proprietary tool ptxas and lowers the PTX code into binary code. We have tried to reproduce this issue on different hosts including the exact RE worker that produced the empty output, but it couldn't be repro'ed.

This issue is tricky since it exits successfully, the empty output will be uploaded to cache, and we only see a failure in the downstream action. Modify the action to check that the output file is not empty and fail if so. This will surface the error early and hopefully the action gets re-run and produces a valid output.

Reviewed By: get9

Differential Revision: D78040423

fbshipit-source-id: 4d233f635042351d3b43fb9641f9d5ae952a6444