Don't exit early on clap error
Summary:
`clap.get_matches_from` calls `exit(0)` for `--help` or `exit(2)` on an error which means it's using the exit code for infra errors to report user errors and not logging invocation records.

Use `try_get_matches_from` in order to not exit early and to log a user error exit code (3).

This adds about 100ms of overhead to `buck2 --help` but it seems reasonable to make the overhead of this match the baseline overhead of almost every other command.

Reviewed By: JakobDegen

Differential Revision: D75979100

fbshipit-source-id: 298d5c84cf54f8bdd220544335b71259232e9d6f