Crash if fails reload.
Summary:
Sometimes I experience an issue with the buck2 run test shell where a module is not reloaded properly. I always loose 5 minutes of my life putting debug instructions there and there till I realize there is sth wrong with the shell.

In this diff we make it crash if the reloading fail. Not sure if it will catch those issues, but might help.

Reviewed By: jcpetruzza

Differential Revision: D78008617

fbshipit-source-id: 65cfd907d78698919f573f4c7ac3e26bdc7c1e9b