Improve manifest-module action
Summary:
When running larger targets there are 17000+ manifest flags passed into this action (i.e. "--module-manifest 1.json --module-manifest 2.json ... --module-manifest 17000.json"). The arg parser in Python will combine these files into one list. Running this locally on my OD, this step took around 10 seconds.

Instead passing 17000 flags, a file with the list of manifest files are passed in. Reading and parsing this is much faster.

This change along with the next diff in the stack brings the execution time of this script from 14 seconds to ~1-2 second locally on my OD.

Reviewed By: zsol, AishwaryaSivaraman

Differential Revision: D78112413

fbshipit-source-id: 3ec4246c325be3e795e859137c536a05a8e2c3ed