python_library: fix bundling late stamped cpp resources
Summary:
We normally don't support packaging resources with `nondebug_runtime_files` in `standalone` package styles. In D38947009 we made an exception to this and applied a heuristic to detect `cpp_binary` targets and their shared lib symlink trees.
That heuristic relies on assumptions that late stamping breaks (namely, that the basename of the artifact follows `__<binary_name>__shared_libs_symlink_tree`, but a late stamped binary's name ends with `-stamped` without changing the symlink tree artifact's name).

This diff adjusts that heuristic, but ultimately we should not be relying on naming conventions at all, as they're fragile.

Reviewed By: christycylee

Differential Revision: D78165266

fbshipit-source-id: c0266c6cd47e16b1c7ee81a4c7eaef77fde6c81b