Allow flushing FileTailers via mut reference
Summary:
This was originally landed as part of D75744746 as a step towards simplifying the lifecycle of FileTailers and to allow moving EventsCtx outside of a tokio context. It's necessary to flush FileTailers without taking ownership in order to hold it in the EventsCtx and flush in more than one place.

That change caused `FileTailers::stop_reading` to not drop stdout/stderr file tailers before flushing which caused buck to hang and tests to timeout. Fixing it by explicitly dropping the tailers in `stop_reading`.

Reviewed By: cjhopman

Differential Revision: D78189474

fbshipit-source-id: 81ebf84090e6605119ec067348f337cc92ef3e35