erlang: don't use dep files for conflict detection
Summary:
This moves header conflict detection from dep files to happen directly on the header files themeselves.

To do that, we need to keep track of private includes in build environment without exposing them to compiler - to do that we store them in a separate "hidden_private_includes" field.

This is necessary since we want to remove the per-file storage of dependencies and instead store them per-application to help with buck2 memory used for storing all those edges in artifact dependency graph.

Reviewed By: TD5

Differential Revision: D78338024

fbshipit-source-id: a30e3249838c7d6e0dfe065ff90fe0cb2d3bb20d