Move event finalizing out of client ctx, again
Summary:
(Originally landed as D75581273, had to revert to back out D75744746)

This is an intermediate step towards finalizing logging at the top of the stack.
This shouldn't capture new errors for existing commands but should cause us to start writing invocation records and errors for a few commands that don't call command_ctx.exec, specifically root and status.
I'm explicitly opting out logging for internal commands that are triggered by other commands (forkserver, internal-test-runner and later, daemon) since if these fail it should be propagated to the triggering command. I'm also not sure if these share the trace id of the parent command which could be an issue.

Reviewed By: cjhopman

Differential Revision: D78189477

fbshipit-source-id: 48354798dbf2b4cd9be31b029cb1ced764129bb9