Add symlink from configuration path when materializing content-based input paths
Summary:
We want to make sure that whenever a content-based path is materialized, we also provide a symlink from the configuration path to the content-based path. So, do that when we materialize input paths for an action.

Note, this doesn't change the action definition, we will still see action dedup take place (if it would have otherwise).

Reviewed By: scottcao

Differential Revision: D78029720

fbshipit-source-id: bdbb20f7ee5a40c73f1274d4cc739836ccb224e4