Introduce GoDistrInfo
Summary:
The standard go distributives are created by `go tool distr` tool and always have the same structure.
It makes sense to create `GoDistrInfo` provider and supply it as a single `exec_dep` for the `go_toolchain` rule.
The `go_distr` rule simplifies buckification of a go distribution and can be used for both local toolchains and `http_archive`.
So we got:
- Reduced number of attributes on `go_toolchain` rule
- Single `exec_dep` for the Go distribution (no more weird [DefaultInfo in exec deps](https://www.internalfb.com/code/fbsource/[cd389a6fded321c01813a1e9ef7a0d560fc25ede]/fbcode/buck2/platform/go_toolchains.bzl?lines=68))
- Simple buckification, no need to create `export_file/command_alias` targets
- Local and `http_archive` toolchains supported.

Reviewed By: echistyakov

Differential Revision: D78278642

fbshipit-source-id: 5502f7610e5d358c4e3e4a0362bb9e45d2e873b1