Swift: add `clang_modulemap_artifacts` to `SwiftCompiledModuleInfo`
Summary:
**Problem**

Clang modulemaps are required by `lldb` for debugging purposes, even with explicitly modules.

To ensure that it's part of the `[debuginfo]` subtarget, we must provide a list of `Artifact`s.

*But* `clang_modulemap_args` is a `cmd_args`, so we cannot provide the modulemap as debuginfo.

**Solution**

Introduce `clang_modulemap_artifacts` on `SwiftCompiledModuleInfo` which will carry the modulemap and required hidden artifacts.

NOTE: The new field is unused, it's only populated with the correct data.

Reviewed By: rmaz

Differential Revision: D78408275

fbshipit-source-id: 85c19b0f7dac33704f5a071b602c2e347cbd98dc