Report non asset consituents for debugging of merge maps
Summary: T229920742 is a crash caused by a merged library becoming a non-asset library as a result of changes to one of the merged targets. When debugging problems like that it would be convenient to be able to look at merge map debug info and find all reasons for a particular merged library being a non-asset library.

Reviewed By: IanChilds

Differential Revision: D78412600

fbshipit-source-id: 4c842ffc6caa9fd3b27cd5558bf1f8026123d210