erlang: use native enforcement for type of test deps
Summary:
We can avoid manual checks on type of dependencies by leveraging native buck2 mechanism of checking provider type.
To do this, we need to introduce a new "mark" provider for things that tests are allowed to depend on.
This will be significant later as we change how dependencies are handled, to do it in a more centralised way.

If an error is produced, this is roughly how it would look like - I think this is clear enough:
>    Attribute requires a dep that provides `ErlangAppOrTestInfo`, but it was not found on `prelude//erlang/applications:common_test (cfg:opt-linux-x86_64-fbcode-platform010-clang19-no-san#29ec9077e7214455)`. Found these providers: DefaultInfo, ErlangAppInfo

Reviewed By: TheGeorge

Differential Revision: D78481055

fbshipit-source-id: e211ab213e3484de588949297b12887e4ff91b71