Add ClientRuntime, again
Summary:
(Originally landed as D76086836, had to revert to back out D75744746)

The next diff finalizes the `EventsCtx` from a non-async context so requires having a tokio runtime at the top of the stack.

Finalizing the event log calls `tokio::io::AsyncWrite::write_all` which panics (on some platforms) if executed by a different tokio runtime than the one that creates it.
We also can't create the runtime earlier since creating a tokio runtime before forking also causes panics (on different platforms).

This holds a reference to a runtime in a `OnceLock` so we can initialize it during runtime and re-use it, or, initialize during finalization if it was never initialized.

Reviewed By: cjhopman

Differential Revision: D78189475

fbshipit-source-id: d4f400c51be663a6aae62090c26eaeaf0a07bdeb