Add struct for map of sketchers
Summary: To report total per-configuration sketches in build report, we need to merge per-configuration sketches of every top-level node, which would involve merging maps of sketchers. This diff adds a `VersionedSketcherMap` struct which holds a map of configuration string to `VersionedSketcher`. This struct supports a `merge` method with other `VersionedSketcherMap` to compute total per-configuration sketch, which we will use in next diff

Reviewed By: IanChilds

Differential Revision: D76098120

fbshipit-source-id: 54b39e93e2f5d70bc5eda96f11cf688a5e1833cc