Change build_metrics to be an Arc
Summary:
Change `build_metrics` field to be an `Arc` in prepartion of displaying `?modifier` syntax in build report.

This is necessay because multiple `TargetLabelWithModifiers` could correspond to a `ConfiguredProvidersLabel` and as a result share an entry in the metrics.  By making it an `Arc` we can now have them all have a copy of the `TargetBuildMetrics` cheaply.

Reviewed By: scottcao

Differential Revision: D78447206

fbshipit-source-id: d81aec4535d454eacf6c1697d4db8368d01ccbc9