Remove inputs from API for registering actions
Summary:
In general, we choose to calculate the `inputs` lazily on demand to save memory, and if an action chooses to pass through inputs instead, then it can just do that manually.

Having this param is a bit confusing because it makes it seem like you should be declaring `inputs` when you register, which isn't true.

Reviewed By: scottcao

Differential Revision: D78331456

fbshipit-source-id: d6ef01e4c9ba199f8af85e1341b814d83f406b25