Move invocation record uploading to top of stack
Summary:
We should try to log an invocation record for every buck command as long as a dependency of logging doesn't fail or we panic or `exit` is called somewhere.

We are logging ~45 million commands a day to buck_results that don't have a corresponding buck2_builds entry, including ~5 million with non-user exit codes. We should try to understand how many of those are from actual infra errors.

We are currently logging ~240 million commands a day to buck2_builds so this could represent a 15% increase in the number of samples, however the increase in storage should be much smaller since:
- The size of these samples should be a fraction of other samples, based on column storage sizes in iData for the few columns this logs (trace id, errors and some metadata) these should be less than 10% of the size of an average sample.
- Some number of wrapper invocations never call buck (--version etc.)

Based on that I would estimate this should increase the size of writes to buck2_builds by <~2% so the impact on retention should not be very significant. We could limit this to only logging errors but it doesn't seem necessary.

Reviewed By: Will-MingLun-Li

Differential Revision: D75581267

fbshipit-source-id: d70f874bab8332a70258fcbad9b91e4b9086f562