Use soft_error quiet setting on the client
Summary:
Presumably this was set because on the client we can't broadcast an instant event, however we still write soft_errors to scribe directly from the client so it shouldn't matter if the daemon doesn't receive an event.

Soft errors should be quiet by default regardless of whether they are sent.

Reviewed By: Will-MingLun-Li

Differential Revision: D78577932

fbshipit-source-id: 0e3ad06e64020c1cc67ac09320ebc412cefb2352