Fix build report tests for python3.10
Summary:
These golden tests fail on my devserver since it's on python3.10 while ODs and SC are still on python3.8.
python3.10 uses absolute paths in error messages by default, and slightly different formatting of ^ characters.

Re-enabling mac tests since these were disabled on macs because of python3.10.

Also stripping "Waiting on" since I'm seeing that on my devserver, not sure if that's related to the python upgrade or why it's not showing up on SC.

Reviewed By: IanChilds

Differential Revision: D77775469

fbshipit-source-id: 9e2ae0db19f59301c9146ef4099d3883f1822107