Update elp.bxl to put extra_includes targets into the deps output
Summary:
Sometimes an erlang application depends on includes from another application, but it cannot be made a dependency because this would cause a dependency loop.

This is sidetracked by having an `extra_includes` field in the `erlang_application` definition for the target, listing dependencies needed only for includes.

For each of these, the buck processing generates a virtual target with the suffix "_includes_only", and pulling out only the includes from the original target.

Up to now, the `elp.bxl` script has not processed these.

This diff adds processing, putting them in the output `extra_includes` field in the bxl result.

Reviewed By: TheGeorge

Differential Revision: D78099637

fbshipit-source-id: 4f69c7598141f17e10deabd53d1c394ba04fcdc8