erlang: share code path
Summary:
This leverages the shared `ErlangDependencyInfo` infrastructure to compue the code path once and share it between different places that want to use it.

In particular we have 4 such places
* `DefaultInfo` of applications - to establish dependency on other application ebin directories and force them to be compiled
* same in tests
* `test_info` file listing code path for the test
* `start_shell.sh` file for applications & tests listing code path for the new VM

With this change all of those places compue the code path in exactly the same way, and because of the anon rule it can be shared if two targets have the same dependencies.

This significantly reduces memory use. Retained memory is now 860MB (down from 1GB), total allocations are 1GB, down from 2GB).

Reviewed By: TheGeorge

Differential Revision: D78665541

fbshipit-source-id: 633907d4723790c60c287d5d2c78550b1c88a029