Use exec_dep() instead of dep() in jvm_common.plugins()
Summary:
Plugins are run on the execution platform (as part of Java/Kotlin compilation), and so they should be built for the execution platform, not for the target platform.

In practice, this hasn't really matter because they're Java binaries, but this does help with sharing cache (there are much fewer execution platforms than target platforms).

Reviewed By: cjhopman

Differential Revision: D78556420

fbshipit-source-id: 2cdea4707b38842067f5ce8f582e0c30167c49b2