Map modifiers when there is configured build error
Summary:
Also map the `ConfiguredProvidersLabel` to `Modifiers` whenever we have a configured build event error.

This code is unfortunate as ideally we would just call the `MapModifier` event just once during the configuring phase, however, I could not find a way to make this happen without a slight hit to performance with an additional call to `get_configured_providers_label`.

Reviewed By: scottcao

Differential Revision: D78624475

fbshipit-source-id: 3b42b3439c82bbc155b60eff29cfbd26a691c05d