Change configured_to_pattern_modifiers value to a BTreeSet
Summary:
Change configured to pattern modifiers mapping value to a `BTreeSet` instead of a `Vec` due to how we are handling `BuildTargetResult` in the test command. Using a `BTreeSet` allows us to more gracefully handle extending `BuildTargetResult`.

Also updated the tests that were incorrectly behaving for the `?modifiers` syntax to now work as intended due to this change.

Reviewed By: scottcao

Differential Revision: D78626239

fbshipit-source-id: 8f95a1cb4dec3bcfdb44095556a564780981f235