Fix asan for macos rust
Summary:
Building rust code with asan enabled on macos introduces linker errors due to rustc looking for `rustc-stable_rt.asan` from the rust sysroot. We don't want that because we do not use the rust sysroot at all.

I had to fake the structure of the sysroot a bit, as rustc is generating a `-L` for the lib directory even with `-Zexternal-clangrt` on macos.

Reviewed By: dtolnay

Differential Revision: D78905823

fbshipit-source-id: 4b00072c4dfd1e0f8d799a4b8ed0ef79f2ba5929