reduce `buck2 test` verbosity
Summary:
Context is the investigation done here T232172653.

`buck2 test` prints the entire list of tests while they are under execution. Although, this is generally not a big issue, things change for CTS Vulkan tests, which are executed into 50 different shard, each having ~11k test cases!

The workflow that executes the tests, listen for the stderr generated by `buck2` and consequently produce an artifact containing it. Such a file is un-necessary big (for some shard of tests can reach also ~3 GB, refer to T232172653) as the full list of tests under execution is repeated for thousands of times.

This Diff proposes a way to solve this issue.

Reviewed By: pringon, mijay

Differential Revision: D79056004

fbshipit-source-id: 3b62c4b707ee8d2e3ff8c8498994cf7a39a7df5e