Content based paths for JavaCD
Summary:
This turns on content-based paths for all JavaCD and KotlinCD actions, except those produced by incremental actions because that is not yet supported (will come in due course).

The overall goal of content-based paths is to produce actions that are identical across different configurations. We can do this if the inputs and outputs are all content-based. Once this lands, I'll take a look at how much duplication we're seeing and track down any other inputs that we need to turn this on for.

For the IDE, we return a configuration-based path in BXL, and we symlink from the configuration-based path to the content-based path whenever we materialize a content-based path. cc cute-jumper

Meta: cc ywett02 in case you want to verify that content-based _inputs_ (deps) to incremental actions do not cause an issue. The snapshot jars are not content-based.

I'll land this Monday in case there are issues.

Reviewed By: NavidQar

Differential Revision: D78665892

fbshipit-source-id: 39f076bcce6ffd9e487a240b0a0fa4006b948e57