Make DiceRejected an environment error
Summary:
DiceRejected can happen due to merge bases changes which should be an environment error.
It can also happen if there is an error initializing the file watcher which potentially could be tier0, but it should be transient, tends to only happen with many concurrent commands and is tracked separately via soft_error so also seems reasonable to exclude from tier0.

Reviewed By: Will-MingLun-Li

Differential Revision: D78751505

fbshipit-source-id: 4673e72889142b5e40e057e7a52f1a436c637e32