Don't report OutdatedEpoch on invalid version
Summary:
`CancellationReason::OutdatedEpoch` is not an expected case and should be a tier0 error.

`CancellationReason::Rejected` is somewhat expected, since it can happen if the merge base changes or the file watcher fails to init. We should only report OutdatedEpoch if the epoch is invalid, otherwise report Rejected if the version is invalid.

Reviewed By: Will-MingLun-Li

Differential Revision: D74363816

fbshipit-source-id: 181a87cc000075c147a44585563d5cdcada9c7d8