Use the default mode on macOS
Summary:
Currently, rust-project uses the default mode on Linux, but a specific modefile on macOS and Windows.

However, the modefile assumes we're working on fbcode, which isn't true on some WhatsApp targets.

Most targets build fine without modefiles these days ("modefile-free fbcode"), and we're already doing this on Linux. Make macOS consistent with Linux.

(I haven't modified Windows as D48049152 and D61225602 both suggest that the modefile is necessary for Windows, so I've left it as-is. The modefile logic was originally added in D47728691.)

Reviewed By: capickett

Differential Revision: D79088314

fbshipit-source-id: 800555a897238fbcf1197003da7b87b8b3b0a6ec