Tag HTTP 403 (Forbidden) errors as environment errors
Summary:
S545337 broke a large number of builds with this error which is currently categorized as user/input. Tag it as environment so we can track it with an SLI (to be created).

This will also cause these to be treated as infra errors in CI and retried, which seems appropriate in this case.

Reviewed By: Will-MingLun-Li

Differential Revision: D79188697

fbshipit-source-id: 310b303421429271929353f347d29de74b6fdc64