adding cxx precompiled headers implementation to prelude
This change updates the precompiled_header rule to compile a PCH and updates both
cxx_library and cxx_executable to compile and link with it. PCH only work for local
builds, and mentioned here (https://github.com/facebook/buck2/issues/877) there are potentially
1.5x-2x speedups on builds for using PCH. Noted here, the current precompiled_header
implementation doesn't really work as expected (https://github.com/facebook/buck2/issues/368)

Some of the decisions I could use feedback on:

 - The current approach maintains backwards compatibility with the old precompiled_header
rule in case anyone was depending on how it was implemented. Upgrading to the new
version requires `compile_pch_file=True` and `pch_language=<any clanguage extension>`.
Alternatively this could be named something like `precompiled_header_v2` without
interfering with the old implementation

 - The PCH and consumer both need to have exactly matching flags. I was
hoping to find a better way to match up the PCH rule with the consumer
because this matching needs to be done by the developer

 - This re-uses the `cxx_library` internals of `cxx_library_parameterized`, which
causes a bunch of shuffling of the input attrs so both the `cxx_library` and
`precompiled_header` rule can receive the same inputs.

 - The PCH rule needs a pch_language to be specified. The cxx internals
infers clanguage by file extension, which we cannot do with a `.pch`
extension so we explicitly pass it in. An alternative could be for the
PCH rule to return a provider with all flavors like how pic/non-pic currently
work and have the consumer pluck out the correct one to use. This
direction looks like it will involve more changes potentially around
`MergedLinkInfo` and I'm not sure if we want to take that approach

example usage:
```bash
cxx_precompiled_header(
    name = "my_pch",
    src = "my_pch.pch",
    deps = [],

    # enabling new PCH flow needs these two fields
    compile_pch_file = True,
    pch_language = ".cpp",
)

cxx_library(
    name = "foo",
    srcs = ["foo.cpp"],
    precompiled_header = ":my_pch",
    deps = [],
)
```