Don't create symlinks to content-based paths when materializing projected artifact inputs
Summary:
Right now, whenever we materialize a content-based path, we add a symlink from the configuration-based path to the content-based path (this is to support things like IDEs where we give them the configuration-based path, since we do not have access to the content-based path).

However, if we have an output `out`, and a projected artifact `out/projected`, then this breaks when we do the following:
- materialize `out`, create a symlink from the configuration-based `out` to the content-based `out`.
- materialize `out/projected`, create a symlink from the configuration-based `out/projected` to the content-based `out/projected`.

If we do that, we end up with `out/projected` not being materialized at all (I don't fully understand why).

For now, I think we can just not bother with these symlinks for projected artifacts (that's what this diff does). I'd appreciate it if someone with a better understanding of symlinks/materialization (scottcao, cjhopman, JakobDegen?) could help me work out how to fix this properly. I would guess what we want to do is to just materialize the non-projected artifact, which I think is what happens under the hood anyway. We have access to that path, but when we're creating symlinks we need to pass an `artifact_value` (https://fburl.com/code/x1wajfns) and the `artifact_value` we have is just for the projected artifact, so I don't think that would work correctly? EDIT: the hack I put together in D79177666 to do just this (pass the `artifact_value` for the projected artifact) does seem to work, if anyone can explain why, that would be much appreciated. Thanks!

Reviewed By: cjhopman

Differential Revision: D79173426

fbshipit-source-id: 9f2503e005cdb5b0bba365e685083d5d8682f9c0