new PACKAGE function
Summary:
This diff creates a new function `test_config_unification_rollout()` in the PACKAGE file, which will be used to control the rollout of the build&test config unification project.

The diff reads the function value and passes it to each `TargetNodeData`. In later diffs, the flag will be passed to tpx.

Reviewed By: scottcao

Differential Revision: D79095463

fbshipit-source-id: adb5ba4aa422ae7252281619be3cda14a5bf781a