Set xplat.available_platforms based on the targets specified
Summary:
xplat has a convention where it will define suffixed targets based on the platforms available. For example, if you have `fbcode//foo:bar` and Apple as a platform, the target `fbcode//foo:barApple` is also available.

The platforms and suffixes are listed here:

https://www.internalfb.com/code/fbsource/[d90626d3b150bc5d3e0b9d6499ff2727eccce326]/tools/build_defs/fb_xplat_suffixed_aliases.bzl?lines=51-57

In D59976108 we set the available platforms to `CXX,FBCODE`. This doesn't work if the user's target doesn't have access to fbcode (e.g. WhatsApp), or if they need additional platforms.

Instead, only offer FBCODE if we're working in an fbcode target, and offer APPLE if we're running on a mac.

Reviewed By: mzlee

Differential Revision: D79090746

fbshipit-source-id: b901c05eb1d8e37e636daef6eaf691ee5170bf04