categorize new EdenErrorType
Summary:
# This diff

Teaches Buck2 to categorize a new type of EdenError, AttributeUnavailable. These errors were previously "GenericError" which was not useful for categorization purposes.

# Context

I'd like to better categorize some errors that we're returning via Thrift. One obvious improvement we can make is differentiate between "GENERIC_ERROR" and missing attributes.

Some attributes such as Blake3, Object IDs, digest hashes, and digest sizes are expected to be unavailable in some cases. The latter 3 are always unavailable when a directory is materialized. Therefore, instead of returning GENERIC_ERROR, we can return "ATTRIBUTE_UNAVAILABLE" which indicates that the requested attribute did not exist for the given file.

ATTRIBUTE_UNAVAILABLE differs from POSIX_ERROR (ENOENT) because the latter may indicate that the tree/file itself was unavailable. In this case, it would be impossible to differentiate between the tree/file being missing and the attribute itself being missing.

ATTRIBUTE_UNAVAILABLE also differs from other unexpected errors, as other errors indicate that the attribute was expected to be available, but it was unable to be fetched due to some other issue.

Reviewed By: Will-MingLun-Li

Differential Revision: D79106393

fbshipit-source-id: 8f9567d14c5d38342163dd59705a96dfbdc48973