inplace: make python interpreter absolute
Summary:
If the python interpreter is a relative path with multiple path components (i.e. of the form `foo/bar`), replace it in the shebang with its absolute path.

These paths are clearly relative to something, and so the resulting par would only be runnable when CWD is that something. That's suboptimal.
THe only exception to that is when the interpreter is a single word, in which case `/usr/bin/env` will do a lookup in `$PATH` first. Arguably this is also a terrible idea, but it's not possible to resolve that without a backwards-incompatible change.

Reviewed By: drinkmorewaterr

Differential Revision: D79552056

fbshipit-source-id: 3ab110b00b31460ff64dfe272a29d0346acacabf