Copy kotlincd_jar outputs to a content-based location
Summary:
Right now, content-based outputs are not supported for incremental actions, and so `kotlincd_jar` outputs cannot be content-based. That means that anything consuming one of those outputs (either the class ABI or the full jar) cannot benefit from action deduplication, because one of its inputs is not content-based.

To workaround that, we can just declare a copy from the configuration-based output to a content-based path. Now, anything consuming the output will get a content-based path, and can benefit from action deduplication.

Reviewed By: NavidQar

Differential Revision: D79553409

fbshipit-source-id: c63c7c5a2ecd0c879d424bd3bcd08c620ce61e30