Add test to show symlinks to content-based outputs are correctly updated
Summary:
We provide a symlink from the configuration-based path to the content-based path for build outputs. This test verifies that the symlink and the content it points are correctly updated and materialized when they change.

Meta: see discussion on https://www.internalfb.com/diff/D79356522?dst_version_fbid=1565787821052952&transaction_fbid=732787196388717

Reviewed By: scottcao

Differential Revision: D79631092

fbshipit-source-id: 6821d3fd25a8bf038931fe7695e525010a931ba9