Fix exported header issue
Summary:
There is a bug where the exported header in the toml file is not working

 {F1980934962}

The root cause is that we are not properly adding the cxx provider due to the if condition check below
```
if values_or(ctx.attrs.cxx_providers, ctx.attrs.native_libs, ctx.attrs.cxx_header_dirs):
     providers.extend(
            get_cxx_providers(
            ...
```
`values_or` checks for not None and it returns the first item that is not None. However, in torch's case, `ctx.attrs.native_libs` is empty list (not None but empty), it gets returned from `values_or` and the if statement is resolved to false because empty is falsy.

Reviewed By: AishwaryaSivaraman

Differential Revision: D79576960

fbshipit-source-id: d96484271adbbc51e083b4a6806a3fa6b8cb637b