Codesigning: include inner codesigning manifests
Summary:
**Problem**

In order to have full codesigning information for an app bundle, we need to expose the codesigning manifests for the transitive bundles.

**Solution**

In this diff, `AppleBundleCodesignManifestTree` is correctly constructed by passing all inner bundles. As everything is already wired up, this results in correct behavior and the transitive manifests end up in the subtarget output of `[codesign-manifest-tree]`.

Reviewed By: rmaz

Differential Revision: D79639296

fbshipit-source-id: 170a80b9e0304dfd519d06c33bd88938df7d085c