directory: Use a custom digester for bundled cell directories
Summary:
This is a little bit more code and a bit ugly, but possibly fixes a bad correctness bug. We have two `DirectoryDigest` traits:

https://www.internalfb.com/code/fbsource/[ada2be654c78]/fbcode/buck2/app/buck2_core/src/directory_digest.rs?lines=18

The `InternableDirectoryDigest` is supposed to be used for types of digests that are strong enough for equality on the digest to imply equality on the directory; the other one is not. The interner itself was correctly requiring the sufficiently strong trait. What should additionally have happened is that the `Eq + Hash` requirements are exclusive to `InternableDirectoryDigest`. Without that, you get implementations like this which are completely incorrect

https://www.internalfb.com/code/fbsource/[ada2be654c78bf3bb506f1265a1d2c87cd1a861e]/fbcode/buck2/app/buck2_directory/src/directory/immutable_directory.rs?lines=128-137

This likely would have lead to incorrect early cutoff on the bundled cell keys, though I don't think it's quite possible to hit that in a real use case, but dangerously close.

I originally attempted to fix this by going through with moving the `Eq + Hash` requirement off of `DirectoryDigest`. I ended up succeeding too (a33c38b74ebad7b409796c752b2df68d1ca241bf), but at a cost of sufficient amounts of trait goop that no one wants to review that

So instead, we do this, and then in the next diff we go back to removing the distinction between the two interner kinds

Reviewed By: IanChilds

Differential Revision: D79611274

fbshipit-source-id: 0c70004b889553488a4cfbc0afb900e4227ba959