Improve unicode support for test listing
Summary: Previously, the testing integrations would crash on non-ascii characters, e.g. in a test name. Here, we add unicode support, so we can at least get as far as test listing, although after that, some integration seems not to like non-ascii so we still get issues, albeit later in the pipeline, so there is further work needed for comprehensive support for running tests with unicode names end-to-end.

Reviewed By: TheGeorge

Differential Revision: D79432033

fbshipit-source-id: 762a37fe78d40a1367e39048e936bbed818244c9