Add a json_file() argument to bxl's cli_args. (#1064)
Summary:
The idea here is to allow passing a path to a json file to a bxl arg.  the clap parser will load the bxl file and return its value as a json object, so the bxl just sees the loaded file content.  This addresses issues with `cli_args.json()` and long json blobs that are too long for an operating system's command line.  It also makes invoking a bxl simpler, since it's generally not feasible to type out actual json content on a command line string.

Pull Request resolved: https://github.com/facebook/buck2/pull/1064

Test Plan:
Imported from GitHub, without a `Test Plan:` line.

Rollback Plan:

Reviewed By: scottcao

Differential Revision: D79751586

Pulled By: Nero5023

fbshipit-source-id: 0082c08798df3ac6c45535eb2568f2ee142c0a2e