parallelism: Cache the value, log the available parallelism
Summary:
This diff is sort-of doing double duty: First, it logs the available parallelism in addition to the configured parallelism. Then, it also caches the available parallelism on a per-process basis. This ensures that the value we log for this field is always the same as the value that the `available_parallelism` function returns, helping put buck into a predictable state.

In the one case where we can properly handle changing parallelisms, make a proper carve-out

Reviewed By: IanChilds

Differential Revision: D79771499

fbshipit-source-id: 349af6cb9ebf7606fc490945e84e2f26bfb45815