parallelism: Add test verifying that num_cpus and std agree
Summary:
I want to switch us over to using the std API instead of num_cpus, since I generally trust std a bit more. To do that safely though, they need to return the same result, so lets add a test that makes sure of that. This test will have to run and pass in all the major environments in which buck runs

I've manually run this test on my devserver, an OD, and a laptop, and it passed in all of those places

This test won't stay around indefinitely - we just need to verify this equivalence at the time of the switch

Reviewed By: IanChilds

Differential Revision: D79771506

fbshipit-source-id: 53edacc742f96ba587158245a5bcdaaf9871566a