Log hybrid scheduling mode
Summary:
It's currently difficult to tell why an action ran locally, depending on the cause, particularly for fallbacks due to RE queuing (added as an outcome in the next diff).

Aside from RE queuing it would help to know if most local actions on CI are running due to local_only or not, and how many are running due to other reasons represented by this outcome.

Note that this may need to be combined with the `execution_kind` of an action (from the last command for an action) to understand the full scheduling outcome, if the hybrid outcome does not specify an executor.

Reviewed By: IanChilds

Differential Revision: D79482329

fbshipit-source-id: 4856929179fad8de3c9807cd50b8c147bba82bab