Report if RE queuing triggered local fallback
Summary:
If RE reports an estimated queue time over some threshold it unblocks local execution to race with the queued RE action.

Currently we detect fallback actions by checking for cases where two actions ran, but one was not cancelled, in this case this works like hybrid execution, if local finishes first the remote action is cancelled which causes this to not be reported as a fallback.

This fixes fallback count so we can track these and adds a separate hybrid_outcome for this case.

Reviewed By: Will-MingLun-Li

Differential Revision: D79482982

fbshipit-source-id: d74bc07e3b1f499216acb6d1d6b04e81eb17938c