buck2.py: run the command from the correct CWD
Summary:
If you invoke `buck2.py` from a directory that has its own `.buckroot` (but doesn't host buck2 itself), that won't work.
This diff runs the outer build (for `buck2`) with a CWD set to the directory hosting `buck2.py`, but then the inner buck2 invocation will be run in the CWD of the `buck2.py` invocation.

Differential Revision: D79886325