do not lose screen recording on timeouts
Summary:
There are some cases were we do not save the screen recording, because we assume that there is an error when there is not. E.g.

https://www.internalfb.com/intern/testinfra/diagnostics/9007199354991152.281475143834126.1753962113/

What happens here is that screenrecord terminates on its own. Then, when we try to terminate it, we cannot. We handle this as a failure, and do not save the screen recording.
With this diff, we detect such cases, and we store the screen recording.

Reviewed By: tagantroy

Differential Revision: D79345725

fbshipit-source-id: f6e07f3f9c04a8fe173c6407b2f52ce3e6f88768