Make cdylib static_pic by default?
Summary:
https://fb.workplace.com/groups/rustatrl/permalink/8516921285098540/

With `advanced_unstable_linking`, we end up linking the rust deps of a cdylib as dynamic rdylibs. There's a point to be made here that `advanced_unstable_linking` is sort-of incompatible with `cdylib`/`staticlib`, as really we should only be statically linking Rust code and not non-Rust deps. However, this appears to work for our codebase and the team that's building this cdylib *does* actually want to export a single statically-linked dylib, even if there are non-Rust deps.

I was also considering enabling `link_style` on `rust_library` to allow specifying:

```
rust_library(
    preferred_linkage = "shared",
    link_style = "static_pic",
)
```

Such that we can produce a statically bundled rust dylib. However that requires a bit more work (testing in D79825388).

#buildall

Reviewed By: dtolnay

Differential Revision: D79827872

fbshipit-source-id: b7d4c27dc5cd7669ada45b3ccaf9ec20312f5973