Declare exit codes with error tag classification
Summary:
Simplifies how exit code is derived from an error and makes handling the same whether derived from an Error or an ErrorReport.

This changes the exit code for errors with no category (see FIXME) to always be UnknownFailure to be consistent, but shouldn't make a difference since we should no longer have errors with no category.

Reviewed By: Will-MingLun-Li

Differential Revision: D79775659

fbshipit-source-id: 01e00b3402a27a29d3b3d08755f80d502468af88