Merge linker.argsfile with final_link_index: take 2
Summary: The original attempt missed the situation where the same linkable group is passed in multiple times, where upon the linker would deduplicate the inputs. For the `pre-flags/post-flags` pair where the corresponding linkables is now missing, we need to ignore the original list of linkables and emit post flags immediately after preflags.

Reviewed By: yulong-zhang-1

Differential Revision: D79909694

fbshipit-source-id: c416e82188af9e75f2c2540c819e93884e410c97