Fix flaky queue estimate cancellation test
Summary:
D79595373 changed the check to cancel actions based on queue estimates from `anticipated_queue_duration.as_secs() >= u64::from(re_queue_threshold)` to `anticipated_queue_duration > re_queue_threshold` to match what we check for fallbacks, but it made this test flaky because sometimes the queue estimate is 0.

I added `BUCK2_TEST_RE_QUEUE_ESTIMATE_S` for the fallback tests because I didn't realize this test always gets a queue estimate from RE due to `--no-remote-cache` and because it's not reusing a daemon (which the fallback tests do).

So it's not strictly necessary to use the env var for either test if we change the check back, but it seems better to be explicit and seems less fragile, especially considering reusing the daemon causes it to fail.

Reviewed By: Will-MingLun-Li

Differential Revision: D79915605

fbshipit-source-id: c683a60f96f30c8fa7b8d24a89953ebf802fd0e6