Augment dep-file.txt instead of kotlin-used-classes.json
Summary:
We have a problem with Kotlin incremental builds and content-based paths, which I wrote a test for in D79802456. This diff is written to solve that problem, which is:
- When we do incremental builds for Kotlin, we calculate the used classes for that invocation, and append those used classes to all the used classes we have from the previous incremental runs.
- With content-based paths, the paths to input ABI`.jar`s can change between incremental invocations. In particular, if an unused `.class` changes, the path to the ABI `.jar` will change.
- When we translate the "used-classes.json" to a dep-file that buck2 can consume, we use a mapping from ABI `.jar` to a representation of the `.class` files. If we cannot find a key for the ABI `.jar` we are interested in, then we just say that we depended upon the whole ABI `.jar`.

It's this final part that is going wrong, because we can end up with an old ABI `.jar` path and a new one, and the mapping that we have only knows how to map the new one to the `.class` representation.

To solve this, instead of appending `used-classes.json`, we can append the `dep-file.txt` that `buck2` consumes. This works because it means we only need to do the mapping for any inputs to the currently running action, which we do have.

So, this diff does that. Unfortunately it ends up being quite big because we have to do it all at once:
- Pass `dep-file.txt` to `KotlincMode` instead of `used-classes.json`, and use that to check that we have the input we need to append to (and if not, we do a full recompilation).
- When we create the previous incremental state, pass `dep-file.txt`.
- When we write the dep file, first read the previous dep file if it exists, and add those entries too.
- ywett02 pointed out that `used-classes.json` is also used to populate `used-jars.json`. I think that output probably shouldn't be used with incremental builds, since it's not entirely precise and the point of that output is to know exactly what you used, but I did the same thing for that one - save off the old version, and read it into memory if we have one when we're populating the new version.

Differential Revision: D79807561

fbshipit-source-id: 578b884866d4eabaf0cffcb7172bf41bc3c096e9