Never preload a bundled runtime's libpython
Summary: I believe this is a bug introduced in D72601654. Bundled runtime means that libpython is already running as part of the PAR entrypoint, preloading it again risks an ODR violation. Linux can somehow understand the ODR issue and merges symbols (I think), but on macOS this leads to assertion failures on module inits.

Reviewed By: christycylee

Differential Revision: D79621786

fbshipit-source-id: c299976a2b359d6888a156a835c9bbd5fbbc05e3