Sort dependencies in rust-project.json
Summary:
Currently the array of deps is unordered, which makes it harder to read the output when debugging.

Instead, sort the deps by crate ID. This ensures that the order of the deps matches the order that the crates are declared in rust-project.json. It also tends to lead to dependencies being roughly ordered 'stdlib, then third-party, then first-party'.

Reviewed By: darichey

Differential Revision: D80086693

fbshipit-source-id: ef650cd51752fb3fc2c0920dd731899094b11749