Remove module_manifest_skeleton from APK based builds
Summary:
This is never passed in from `android_binary_for_automation`
on APK based builds.

It is, however used in AAB based builds.

For APK based builds the voltron postprocessor is the thing that adds
manifests, as part of our migration to AAB and the deprecation of the
voltron apk postprocessor I want to start cleaning up dead code paths

This is also going to be helpful as we integrate Voltron into  WA

Reviewed By: bchang7

Differential Revision: D79778375

fbshipit-source-id: f75c3239422e8d30cb10b2043eba692bc852a489