Rework how we materialize content-based outputs for local actions
Summary:
I added a large inline comment to try to explain what's going on here, but for some extra context on why this diff is necessary, there's a bug in the current implementation where the following happens (I believe, it's really hard to repro).
- A local action and a remote action that write the same content-based output finish at approximately the same time.
- The local action checks if an artifact is declared at the content-based path, and that returns `False`.
- The local action copies the output from the output path to the content-based path.
- The remote action declares a CAS output at the content-based path. This causes the materializer to clean the output path.
- The local action calls `declare_existing` on the content-based path, but there is no longer any content there.

This fixes the issue by moving the copy into the materializer. Now, we declare_existing on the "placeholder" output, and then declare a copy from the placeholder output to the content-based output. If the remote action is declaring a CAS output at the "same time", we will just choose whichever method was declared first, since the `ArtifactValue` must be the same (it's a content-based path). Then when we call `materialize`, we will either do the `copy` or download the output from CAS.

Reviewed By: scottcao

Differential Revision: D80080658

fbshipit-source-id: f207aee9a0ef79f64d9bedd0cff6346ff4f3c664