Add logging of per-test external config
Summary:
We take the config derived from the buck CLI and update it with per target information.  This is needed, e.g., for per target modifiers and per-target default modes.

This means we don't have the external config until very late in the run.  This is one of the earliest places we can log the external config with the per-target modifications.  It would be useful if we could log it elsewhere, e.g. on upload of the results, too, but that's a later problem.

This logging unblocks e2e tests for per-target configurations.

Reviewed By: SkyterX

Differential Revision: D80125482

fbshipit-source-id: 393fa117bcdb504ddf7c5dd015b5c6298e8f5c23