Add a 'test_device_type' for apple_test targets
Summary:
This stack adds support for iPad testing.

The buck build definitions for Apple testing were created for iOS with a few checks added to see if a test is for macOS. We'd like to support new platforms for testing (ie. Catalyst and iPad), but these don't map cleanly to the SDK type.

This diff adds a `test_device_type` parameter to `apple_test` targets, so that a test target can override the device type picked based on the SDK.

Reviewed By: cpepin

Differential Revision: D80034277

fbshipit-source-id: 0578647a254991b3cf4321f9d1117e7d22489bd4