Allow clippy::uninlined_format_args
Summary:
This lint is failing buck2-oss jobs on revisions that used to pass, I can repro it with codesync but not directly in the repo, despite the fact that codesync does not modify any relevant code.

It's unclear to me how a lint could be flaky, presumably it's not the lint itself, but in any case this doesn't seem useful and it's disabled elsewhere in the repo.

Reviewed By: Will-MingLun-Li

Differential Revision: D80303763

fbshipit-source-id: 23f8b83422bacd8efca2953839ab827c5114ae4e