Use TSets for most source_only_abis
Summary: TSets are generally more efficient than traversals, and result in fewer edges in the DICE graph. Calculating the source-only ABI deps is a bit complicated because they can be specified in two ways, which is why they didn't use TSets. However, it's pretty rare that targets ever specify `source_only_abi_deps`, and in the case where there are no `source_only_abi_deps` specified, we can just use a traversal on the `CompilingDepsTSet` that skips over anything that is not specified with `required_for_source_only_abi`.

Reviewed By: NavidQar

Differential Revision: D80320301

fbshipit-source-id: 72e88467c965fd8a19b567be163e6f9be4fd3891