oss: fallback to default watcher if edenfs set
Summary:
Running oss tests locally with the edenfs watcher set in `~/.buckconfig.d` throws an "invalid file watcher" error, some newer examples explicitly set a file watcher to get around this but it's not set everywhere.

It's easier to treat "edenfs" the same as not setting a file watcher for OSS instead of having to always set a specific file watcher in every OSS buckconfig we might need to test.

Reviewed By: Will-MingLun-Li

Differential Revision: D80279111

fbshipit-source-id: f334f938cb624331537fdbbfea9a5f663d04cecf