Expose the test_host_app_bundle through the xcode-data subtarget for use in project generation
Summary:
# Context
When a unit test explicitly specifies a TEST_HOST in Xcode, such as [ios_integration_test](https://www.internalfb.com/code/fbsource/[a9ea7bed42f8]/tools/build_defs/apple/ios_integration_test.bzl?lines=4) rules do, we need to set the TEST_HOST to point specifically to the binary contained in the test host app bundle. We already pass the path to that binary through to the project generator.

We now copy output from buck-out into Xcode's derived data dir to avoid stale builds due to config hash mismatches, so now this TEST_HOST path needs to point to the path in Xcode's BUILT_PRODUCTS_DIR, not buck-out. But since the path to the binary within the app bundle isn't consistent between platforms, we need to calculate the relative path of the binary to the bundle.

# This Diff
Pass the test host app bundle path down through the xcode-data subtarget and expose it in the project generator. Update the relevant tests. We'll use this in conjunction with the binary path to identify the bundle-relative path of the binary.

Differential Revision: D80317704

fbshipit-source-id: 5f59a77275b20039c7a242ab54d7c28168ae93c0