Don't use has_artifact_at for content-based local actions
Summary:
Not needed, we can just let the materializer take care of this.

From scottcao's comment on D80080658: "I'm not the biggest fan of using has_artifact_at since it requires taking a snapshot of materializer that could become stale in theory, doesn't acount for things like symlinks, and was introduced originally has a bit of hack for bundled prelude. In theory, you just need to declare things to materializer. If same artifact value was declared at the same path, then that should be a noop. If a different artifact value was declared at the same content-based path, then materializer should throw an error."

Reviewed By: scottcao

Differential Revision: D80255987

fbshipit-source-id: ee7cb07b818e041d77580abd0940416a3164cae2