Return all files that match a qualified path to enable py vs pyi preference
Summary:
Currently, `pyrefly.bxl` returns its output as:
`dict[target, {"deps": list[target], "srcs": dict[qualified_path, path]}]`. This is great if we only need the `.py` or `.pyi` file, and can select for it within the BXL file due to its simplicity (less work to do in Pyrefly), but it won't work well for IDE.

This diff changes the `srcs` part to return a `dict[qualified_path, list[path]]` instead, which enables us to choose which file to use when performing IDE operations or type checks.

Reviewed By: grievejia

Differential Revision: D80498571

fbshipit-source-id: 57edb500457b7ae552fb8a898051ed62d596a6bf