Add missing references for dwp actions
Summary:
We are seeing flaky failures in an fbpkg that relies on this code path. I suspect that this is because `run_dwp_action` has dependencies on non-registered `.opt.o` files, which we are failing to find when we get cache hits on earlier, local-only actions after late stamping is enabled.

This diff addresses the issue by explicitly adding all previously declared hidden deps for the final link to `referenced_objects`. Feedback is welcome in case I missed anything or if there is a better way to do this.

Reviewed By: christycylee

Differential Revision: D80044477

fbshipit-source-id: 959f22973b0149b59491f51e32845a9a5d780251