Add toolchain support for `--push-state`/`--pop-state` linker flags
Summary:
Many linkers provide "sandboxing" flags (e.g. `--push-state`/`--pop-state`,
https://man.archlinux.org/man/extra/lld/ld.lld.1.en#push-state),
which can be used to isolate some other flags (e.g. `--as-needed`) to
a subset of the libraries on the link line.

This diff adds opt-in support for toolchains to use these flags to wrap
the exported linker flags coming from rules/targets.

Reviewed By: athmasagar

Differential Revision: D80538408

fbshipit-source-id: e169f2632c513bfb6876fa796b816400a8984db7