Account for content-based paths in audit output
Summary:
If we have a content-based path, we don't know the configuration of the action that produced it, and in any case, there could be multiple such actions. Therefore, we just return the unconfigured target label.

To be honest, the whole thing about finding which action created the output seems...overengineered? I'm not actually sure why anyone would need that information, and in my quick grep, I didn't find any usages for it. Not sure if it's worth removing that as a follow-up or not.

Reviewed By: Will-MingLun-Li

Differential Revision: D80617649

fbshipit-source-id: 61e06510b8c13d66cd15bccfc4ea835721957cf5