Use flavor flags both for pic and optimisation compilation
Summary:
This change completely migrates `cxx_library` to constructing artefacts via compile flavors.

This is large scale change that touches `pic` flags for libraries. The good thing about potential breaks is they are build time failures and very clear. So I expect everything to go smoothly but if it doesn't, we'll know and won't break in runtime.

Note: this changes the order of the flags a bit internally. This is essentially a noop, but makes it inconsistent with current value so busts buck cache.

Reviewed By: yahayaohinoyi

Differential Revision: D78023438

fbshipit-source-id: 43cd0d72e13d578705c312f3b9d043c2f7e81eec