Fix build count for incremental failed commands
Summary:
D73794593 introduced a bug that causes min_build_count_since_rebase to not be set for commands that fail which causes incremental failed builds to be counted as cold builds.

The bug is that BuildResult is only set if a command succeeds, meaning we can't use it to determine if a command is one of the command types that run builds and should report `min_build_count_since_rebase`.

Reviewed By: IanChilds

Differential Revision: D80675398

fbshipit-source-id: 92bee70c8d810bb631f4b0cb02544a4c7e3abaa6