Make argsfiles subtargets show an output path
Summary:
`--show-output` only shows an output path if there is a single output.

When build `argsfiles` targets, we don't get the path of their output easily.

To fix this, make the default output the summary, and include the argsfiles as part of other outputs.

We also need to update the `check_argsfile_contents.py` which no longer receives a list of the (relative path) argsfiles, to instead compute the relative path and read the files.

Reviewed By: rmaz

Differential Revision: D80736907

fbshipit-source-id: b46009a7c3ccc2c62878298f3cc13c16b42e73fd