Add `exit-when` flag to Buck
Summary:
See title. This allows the user to specify that a command should fail based on certain conditions. It currently has 3 options:
- `=never` (default) - Runs this command normally. If another command is running with the same state, this command will run concurrently. If another command is running with a different state, this command will queue until the other command exits.
- `=differentstate` - Fails this command if another command is already running with a different state. Otherwise, runs this command normally.
- `=notidle` - Fails this command if the daemon is not idle (regardless of whether the state is the same of not).

The `--exit-when=differentstate` option is functionally equivalent to the existing flag `--exit-when-different-state`. The latter flag will eventually be deprecated.

One key benefit here is that tooling/background processes can give deference to user commands by setting the --exit-when=notidle flag for their commands. Without this a background process could start a command while a user-triggered build is running, potentially causing the user command to compete with tooling for the same resources on the user's machine.

*Interplay with --preemptible flag on other commands*
Critically, the `--exit-when=notidle` option does NOT preempt other commands that are already running with the --preemptible=always or --preemptible=ondifferentstate options. If any build is running (preemptible or otherwise), then the `--exit-when=notidle` build will not start.

Reviewed By: JakobDegen

Differential Revision: D78699268

fbshipit-source-id: 9862b6342f7133195590c9933e4b033660ccc54a