Add an `executable_name` parameter
Summary:
Some tools rely on binaries having particular names. For example, LLVM's
lit test runner expects the compiler to be named `clang`, the ELF linker
to be named `ld.lld`, etc. If we need to wrap these binaries within a
`command_alias`, that tooling fails. Add an `executable_name` paramater
to `command_alias`, which matches the existing `executable_name`
parameter on `cxx_binary`, to let us specify a custom name for the
command alias trampoline script and avoid this issue.

Reviewed By: JakobDegen

Differential Revision: D80831249

fbshipit-source-id: d6e4818bf3c62b3d905c35612ad45ad80ffcb012