Make rust_cxx_bridge use content-based genrules
Summary:
Since those end up in cxx compilations, which we want to dedupe.

The updated tests are for coverage and simply test that the generated files in question are present in the coverage file, so it should not matter that the generated path has changed (unlike e.g. if it was testing for some IDE consistency as we had with Apple rules before).

Reviewed By: dtolnay, capickett

Differential Revision: D80903374

fbshipit-source-id: 1817eb2a783f51d09e5efbadcb34e4b6b73b43c2