Move memory limit checking into memory tracker
Summary:
We should report events when we go above and below the memory limit to validate how well we are triggering on OOMs and it doesn't make sense to do that in `LocalActionsThrottle` since that will fire an event for each action, instead of once per state change.

We're also planning to use the memory tracker to track pressure stall information and it makes sense for that to be configured inside the memory tracker, which should reuse similar logic.

Reviewed By: Will-MingLun-Li

Differential Revision: D80500904

fbshipit-source-id: 738ee0b2c5bbea8306f9716b84e82351285e7579