Separate MemoryTracker from state sender/receiver
Summary:
Currently `MemoryTracker` is immutable because it's accessed via an Arc from the throttler and inside the memory tracker loop.

We only need to share the sender, not the whole MemoryTracker, so we should only share that.

Needed to store mutable state related to action cgroups in a later diff.

Reviewed By: Will-MingLun-Li

Differential Revision: D80582550

fbshipit-source-id: 23bbe7ef3ac9f8deb056e71b91d38e94abe79210