Add per command memory tracker reporter
Summary:
We should report memory pressure events to the invocation recorder and to scribe. I experimented with sending events directly from the tracker but it makes sense to do this per command because we should check the state when commands start and any spans should be per command.

This adds debug statements as a placeholder, replaced in next diff.

Reviewed By: Will-MingLun-Li

Differential Revision: D80833076

fbshipit-source-id: 448e5da52e9211160e56a7e957372d7ce30d4adf