Get absolute path from BXL
Summary:
We need to make everything an absolute path to ensure Pyrefly works well, and that it's runnable from locations other than the directory that's the root of the repository.

To do this, I'm going to return the buck root from the BXL. I'm doing this in the query, since I think it's preferable right now, especially if the repo root won't be the same on all systems. Then, we can pass in the config path as CWD to the BXL, and this should get us in a spot where we can rewrite the paths.

# This diff stack...
works on emulating Buck's (and other build systems') file mappings into something that Pyrefly can model and work with.

We want to do this because:
- it will help Pyrefly work with build systems better in OSS
  - `pyrefly buck-check` exists, but assumes a specific model of how Pyrefly will be called and interact with build systems, which isn't necessarily how other build systems will work
- it will enable Pyrefly to work with build systems in the IDE

## The idea...
Model a virtual filesystem in Pyrefly. We can do this by taking from what Neil set up with in `pyrefly.bxl`, which is a way to query Buck for a mapping of `{<target>: {"srcs": {<module_name>: <module_path>}, "deps": <target>}`. This approach appropriately models the build system by keying everything on `Targets`, where *only* targets are unique, and files can appear as sources in multiple targets, and module names are also only unique *within a target*.

Given this information, we can have the sourcedb we provide to a `ConfigFile` able to understand this, and correctly understand that, given a module name, file, and target, which file the import is referring to.

With that information, it's only a matter of knowing when to re-query the sourcedb, and we should be able to replicate build systems in Pyrefly!

Reviewed By: grievejia

Differential Revision: D80854253

fbshipit-source-id: 44d2070d1cb84656d291b3391a10d51c453a7925