Allow lists of strings for erl_opts and emu_flags in erlang_toolchain rule
Summary:
# Context
When defining an erlang toolchain, the user can provide the compiler options (`erl_opts`) and VM options (`emu_flags`) to pass when using the toolchain.

# Problem
At the moment, these are specified as strings by the user, and converted to a list of string internally by splitting at spaces. A string is convenient when `emu_flags` and `erl_opts` are defined in buck2 config (bcfg files, etc)

For specifying these attrs manually, a list would be preferrable. In fact, if one wants to use constraints/modifiers to configure these options, then using a list is important, as `select()` can be "added" to lists, while combining with strings becomes cumbersome quickly.

# This diff
We make it so that lists of strings can also be used to configure the toolchain, while still accepting strings, for backwards compatibility

Reviewed By: TD5

Differential Revision: D81150510

fbshipit-source-id: 00f659c5d96cd3b863c33207424979bebe9f2a42