Fork Server Integration with cgroup pool
Summary:
Fork Server Integration. In command stream,  when we get the exit event we call the on exist to drop the cgroup to release it back to pool.

But since our kernel version is at 6.9, right now we cannot reset cgroup v2 memory.peak (which is available when kernel version >= 6.12). If we use cgroup pool, the memory peak from minipref is not correct. We need better way to test it.

Reviewed By: christolliday

Differential Revision: D80142186

fbshipit-source-id: 2517253d7a19398bae09329be1fbfaec2bbf0432