Use UUID for action cgroup unit names
Summary: Action digests are not unique within the same command due to content based hashing. It's also possible systemd units might not be cleaned up between buck commands in which case they should be globally unique. If that means we are leaking cgroups, it should be resolved by cgroup pools later.

Reviewed By: JakobDegen

Differential Revision: D81338115

fbshipit-source-id: e162ee1db9071ee3fbd495ac7b3b65cdc2f7669d